Move Exception to new namespace
