+++ src/Barcode/Exception/BarcodeValidationException.php
@@ -6,11 +6,11 @@
  * @license   https://github.com/laminas/laminas-barcode/blob/master/LICENSE.md New BSD License
  */
 
-namespace Laminas\Barcode\Object\Exception;
+namespace Laminas\Barcode\Barcode\Exception;
 
 /**
  * Exception for Laminas\Barcode component.
  */
-class ExtensionNotLoadedException extends RuntimeException
+class BarcodeValidationException extends InvalidArgumentException
 {
 }
