+++ src/Barcode/Exception/RuntimeException.php
@@ -6,7 +6,7 @@
  * @license   https://github.com/laminas/laminas-barcode/blob/master/LICENSE.md New BSD License
  */
 
-namespace Laminas\Barcode\Object\Exception;
+namespace Laminas\Barcode\Barcode\Exception;
 
 use Laminas\Barcode\Exception;
